separate config logic
