+++ .gitignore
@@ -7,7 +7,7 @@
 node_modules/
 dist/
 
-config.json
+.config.json
 
 # Compiled Java class files
 *.class
